style(user-srv): user-srv模块代码优化

+++ service-share/user-srv/src/main/java/com/lawu/eshop/user/srv/service/impl/MerchantStoreInfoServiceImpl.java
@@ -488,9 +488,9 @@ public class MerchantStoreInfoServiceImpl implements MerchantStoreInfoService {
 			// 填写身份证用户需要交保证金
 			isShow = false;
 			//修改店铺状态
-			storeDO.setStatus(MerchantStatusEnum.MERCHANT_STATUS_NOT_MONEY.val.byteValue());
+			storeDO.setStatus(MerchantStatusEnum.MERCHANT_STATUS_NOT_MONEY.val);
 		} else {
-			storeDO.setStatus(MerchantStatusEnum.MERCHANT_STATUS_UNCHECK.val.byteValue());
+			storeDO.setStatus(MerchantStatusEnum.MERCHANT_STATUS_UNCHECK.val);
 		}
 		merchantStoreDOMapper.updateByPrimaryKeySelective(storeDO);
 		merchantStoreAuditDO.setIsShow(isShow);
